requirements.txt
google-generativeai
langgraph
numpy
arxiv
pdfminer.six
requests
certifi
scikit-learn
streamlit
pandas
plotly
graphviz